+++ tags/inttag.h
@@ -7,22 +7,22 @@ class IntTag : public Tag {
     public:
         IntTag(std::string name, int32_t data = 0) : Tag(name){ this->data = data; }
         void NbtPrintData() override {
-            std::cout << "(Int) " << GetName() << ": " << (int)data << std::endl;
+            std::cout << "(Int) " << GetName() << ": " << static_cast<int32_t>(data) << std::endl;
         }
         uint8_t GetTagId() override {
-            return (uint8_t)TAG_INT;
+            return  static_cast<uint8_t>(TAG_INT);
         }
         void Write(std::ostringstream& stream, bool primary = true) override {
             if (primary) {
                 WriteHeader(stream);
             }
-            uint32_t writtenData = Swap32(data);
+            uint32_t writtenData = Swap32(static_cast<uint32_t>(data));
             stream.write(reinterpret_cast<const char*>(&writtenData), sizeof(writtenData));
         }
         void Read(std::istringstream& stream) override {
             uint32_t rawData;
             stream.read(reinterpret_cast<char*>(&rawData), sizeof(rawData));  // Read raw bytes for integer
-            data = Swap32(rawData);
+            data = static_cast<int32_t>(Swap32(rawData));
         }
         int32_t GetData() {
             return data;
